chore(rbo,rules): separate the rule from common rbo code
