+++ dbcon/mysql/CMakeLists.txt
@@ -43,6 +43,7 @@ set(libcalmysql_SRCS
     is_columnstore_extents.cpp
     columnstore_dataload.cpp
     rulebased_optimizer.cpp
+    rbo_apply_parallel_ces.cpp
 )
 
 set_source_files_properties(ha_mcs.cpp PROPERTIES COMPILE_FLAGS "-fno-implicit-templates")
